# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tow truck marketplace API — orders, drivers, payments, payouts, admin.",
        "title": "EVIK Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns orders filtered by status. Clients see their own orders; drivers see searching orders or their own assigned orders; admins see all.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "enum": [
                            "searching",
                            "accepted",
                            "arrived",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "type": "string",
                        "description": "Order status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Max items (1..100, default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.OrderListResponse"
                        }
                    },
                    "400": {
                        "description": "invalid limit",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden role or driver gate failed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a tow truck order. Pickup and dropoff coordinates must be inside an active service area.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.SingleOrderResponse"
                        }
                    },
                    "400": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "service area not allowed or wrong role",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{orderID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.SingleOrderResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "order not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CoordinateDTO": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number",
                    "example": 55.7522
                },
                "lng": {
                    "type": "number",
                    "example": 37.6156
                }
            }
        },
...
        "http.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "auto_dispatch": {
                    "type": "boolean",
                    "example": false
                },
                "dropoff_lat": {
                    "type": "number",
                    "example": 55.76
                },
                "dropoff_lng": {
                    "type": "number",
                    "example": 37.62
                },
                "payment_method": {
                    "type": "string",
                    "enum": [
                        "cash",
                        "card"
                    ],
                    "example": "card"
                },
                "pickup_lat": {
                    "type": "number",
                    "example": 55.7522
                },
                "pickup_lng": {
                    "type": "number",
                    "example": 37.6156
                },
                "tow_truck_type": {
                    "type": "string",
                    "enum": [
                        "winch",
                        "platform",
                        "manipulator"
                    ],
                    "example": "winch"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "http.OrderDTO": {
            "type": "object",
            "properties": {
                "cancelled_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "string"
                },
                "dropoff": {
                    "$ref": "#/definitions/http.CoordinateDTO"
                },
                "dropoff_lat": {
                    "type": "number"
                },
                "dropoff_lng": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "pickup": {
                    "$ref": "#/definitions/http.CoordinateDTO"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lng": {
                    "type": "number"
                },
                "status": {
                    "type": "string",
                    "example": "searching"
                },
                "tow_truck_type": {
                    "type": "string",
                    "example": "winch"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "http.OrderListResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.OrderDTO"
                    }
                }
            }
        },
...
        "http.SingleOrderResponse": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/http.OrderDTO"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}